+++ lib/string_contains.dart
@@ -1,3 +1,4 @@
 library string_contains;
 
 export 'src/string_contains.dart';
+export 'src/clean_widget.dart';