+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 * Extension on [String] that provides a methods to check, if a string contains specific words/phrases/Regex/etc.
 
+## 0.0.2
+* Extension on String that provides a methods to check, if a string contains any bad-words/url/link/email/phone number/hashtag/mention and more others.
+* Added CleanWidget
+* Added methods to check if a string contains a hashtag, a mention.